+++ CHANGELOG.md
@@ -10,6 +10,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- `Executor` API: `.onWorkerThread` renamed to `onNetworkThread`, added: `onV8Thread`
+
+
 ## [0.4.0] - 2018-10-29
 
 ### Added