+++ CHANGELOG.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Changed
-- `Executor` API: `.onWorkerThread` renamed to `onNetworkThread`, added: `onV8Thread`
+- `Executor` API: `.onWorkerThread` renamed to `.onNetworkThread`, added: `.onV8Thread`
+- removed the requirement for `singleTask` launch mode on sign in
 
 
 ## [0.4.0] - 2018-10-29