+++ CHANGELOG.md
@@ -10,10 +10,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unrelease
+## [0.4.2] - 2018-12-08
 
 ### Changed
 - fixed issues #117 and #118 for putFile and getFile
+- Uses custom tabs instead of normal browser view for auth
 
 ## [0.4.1] - 2018-11-05